Move Modal into api.interactions.modals
